modules/camera_module/requirements.txt
azure-iot-device~=2.0.0
opencv-python
requests
flask